# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: EcoLearn API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /students/{id}:
    get:
      operationId: getStudent
      tags: [students]
      summary: Get student profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Student profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentProfile"
        "404":
          description: Not found
    patch:
      operationId: updateStudent
      tags: [students]
      summary: Update student profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentUpdate"
      responses:
        "200":
          description: Updated student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"

  # ─── Lessons ─────────────────────────────────────────────
...
  /quizzes:
    get:
      operationId: listQuizzes
      tags: [quizzes]
      summary: List all quizzes
      parameters:
        - name: topic
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of quizzes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Quiz"
    post:
      operationId: createQuiz
      tags: [quizzes]
      summary: Create a quiz (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizInput"
      responses:
        "201":
          description: Created quiz
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quiz"
  /quizzes/{id}:
    get:
      operationId: getQuiz
      tags: [quizzes]
      summary: Get quiz with all questions
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Quiz with questions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizWithQuestions"
        "404":
          description: Not found
    patch:
      operationId: updateQuiz
      tags: [quizzes]
      summary: Update a quiz (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizUpdate"
      responses:
        "200":
          description: Updated quiz
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quiz"
    delete:
      operationId: deleteQuiz
      tags: [quizzes]
      summary: Delete a quiz (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /quizzes/{id}/questions:
    post:
      operationId: addQuizQuestion
      tags: [quizzes]
      summary: Add a question to a quiz (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizQuestionInput"
      responses:
        "201":
          description: Created question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizQuestion"
  /quizzes/{id}/attempt:
    post:
      operationId: submitQuizAttempt
      tags: [quizzes]
      summary: Submit a quiz attempt
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizAttemptInput"
      responses:
        "200":
          description: Quiz result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizResult"
  /questions/{id}:
    patch:
      operationId: updateQuizQuestion
      tags: [quizzes]
      summary: Update a quiz question (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizQuestionUpdate"
      responses:
        "200":
          description: Updated question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizQuestion"
    delete:
      operationId: deleteQuizQuestion
      tags: [quizzes]
      summary: Delete a quiz question (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Student:
      type: object
      required: [id, name, email, ecoPoints, school, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        ecoPoints:
          type: integer
        school:
          type: string
        grade:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    StudentProfile:
      type: object
      required: [id, name, email, ecoPoints, school, badges, recentActivity]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        ecoPoints:
          type: integer
        school:
          type: string
        grade:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        badges:
          type: array
          items:
            $ref: "#/components/schemas/Badge"
        recentActivity:
          type: array
          items:
            $ref: "#/components/schemas/ActivityItem"
        createdAt:
          type: string
          format: date-time
    StudentUpdate:
      type: object
      properties:
        name:
          type: string
        school:
          type: string
        grade:
          type: string
        avatarUrl:
          type: string

    # ─── Dashboard ───────────────────────────────────────────
...
    ActivityItem:
      type: object
      required: [id, type, description, points, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [quiz, challenge, badge]
        description:
          type: string
        points:
          type: integer
        createdAt:
          type: string
          format: date-time
...
    Quiz:
      type: object
      required: [id, title, topic, pointsPerQuestion, questionCount, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        topic:
          type: string
          enum: [climate_change, waste_management, water_conservation, energy_saving, biodiversity]
        description:
          type: ["string", "null"]
        pointsPerQuestion:
          type: integer
        questionCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    QuizInput:
      type: object
      required: [title, topic, pointsPerQuestion]
      properties:
        title:
          type: string
        topic:
          type: string
          enum: [climate_change, waste_management, water_conservation, energy_saving, biodiversity]
        description:
          type: string
        pointsPerQuestion:
          type: integer
    QuizUpdate:
      type: object
      properties:
        title:
          type: string
        topic:
          type: string
          enum: [climate_change, waste_management, water_conservation, energy_saving, biodiversity]
        description:
          type: string
        pointsPerQuestion:
          type: integer
    QuizWithQuestions:
      type: object
      required: [id, title, topic, pointsPerQuestion, questions, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        topic:
          type: string
        description:
          type: ["string", "null"]
        pointsPerQuestion:
          type: integer
        questions:
          type: array
          items:
            $ref: "#/components/schemas/QuizQuestion"
        createdAt:
          type: string
          format: date-time
    QuizQuestion:
      type: object
      required: [id, quizId, questionText, options, correctOption, order]
      properties:
        id:
          type: integer
        quizId:
          type: integer
        questionText:
          type: string
        options:
          type: array
          items:
            type: string
        correctOption:
          type: integer
        explanation:
          type: ["string", "null"]
        order:
          type: integer
    QuizQuestionInput:
      type: object
      required: [questionText, options, correctOption]
      properties:
        questionText:
          type: string
        options:
          type: array
          items:
            type: string
          minItems: 2
        correctOption:
          type: integer
        explanation:
          type: string
        order:
          type: integer
    QuizQuestionUpdate:
      type: object
      properties:
        questionText:
          type: string
        options:
          type: array
          items:
            type: string
        correctOption:
          type: integer
        explanation:
          type: string
        order:
          type: integer
    QuizAttemptInput:
      type: object
      required: [answers]
      properties:
        answers:
          type: array
          items:
            $ref: "#/components/schemas/QuizAnswer"
    QuizAnswer:
      type: object
      required: [questionId, selectedOption]
      properties:
        questionId:
          type: integer
        selectedOption:
          type: integer
    QuizResult:
      type: object
      required: [attemptId, score, totalQuestions, correctAnswers, pointsEarned, feedback]
      properties:
        attemptId:
          type: integer
        score:
          type: integer
        totalQuestions:
          type: integer
        correctAnswers:
          type: integer
        pointsEarned:
          type: integer
        newBadges:
          type: array
          items:
            $ref: "#/components/schemas/Badge"
        feedback:
          type: array
          items:
            $ref: "#/components/schemas/QuestionFeedback"
    QuestionFeedback:
      type: object
      required: [questionId, correct, correctOption]
      properties:
        questionId:
          type: integer
        correct:
          type: boolean
        correctOption:
          type: integer
        explanation:
          type: ["string", "null"]
...
    Badge:
      type: object
      required: [id, name, description, icon, pointsRequired]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        icon:
          type: string
        pointsRequired:
          type: integer
        earnedAt:
          type: ["string", "null"]